api/requirements.txt
redis
flask
flask_cors
requests
logger
jwcrypto
pyjwt